# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: Mobile Travel App API
  description: |
    Comprehensive REST API for Mobile Travel App Backend System

    ## Features
    - **User Management**: Admins, Customers, and Drivers
    - **Coin System**: Top-up requests and transaction tracking
    - **Route & Vehicle Management**: Travel routes and vehicle fleet
    - **Schedule Management**: Trip scheduling with conflict detection
    - **Ticket Booking**: Passenger booking with seat selection
    - **Travel Documents**: Digital travel permits with PDF generation
    - **Driver Panel**: Real-time trip tracking and passenger management
    - **Trip Logs**: Complete audit trail of all trips

    ## Authentication
    All endpoints require JWT Bearer token authentication unless marked as public.

    ## Coin System
    - Ticket booking: 10,000 coins per passenger
    - Travel document: 10,000 coins per document

  version: 1.0.0
  contact:
    name: API Support
    email: support@travelapp.com
    url: https://github.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development
  - url: https://api.travelapp.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    VehicleType:
      type: string
      enum: [EKSEKUTIF, REGULAR]
    VehicleStatus:
      type: string
      enum: [AVAILABLE, IN_USE, MAINTENANCE, RETIRED]

    # Schedule & Ticket
...
paths:
...
  /vehicles:
    post:
      tags: [Vehicles]
      summary: Add vehicle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vehicleNumber, type, capacity]
              properties:
                vehicleNumber:
                  type: string
                  pattern: '^[A-Z]{1,2}\s?\d{1,4}\s?[A-Z]{1,3}$'
                  example: B 1234 ABC
                type:
                  $ref: '#/components/schemas/VehicleType'
                brand:
                  type: string
                  example: Mercedes-Benz
                model:
                  type: string
                  example: Sprinter
                capacity:
                  type: integer
                  minimum: 1
                  example: 16
      responses:
        '201':
          description: Vehicle added

    get:
      tags: [Vehicles]
      summary: List vehicles
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/VehicleType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VehicleStatus'
      responses:
        '200':
          description: List of vehicles
  /vehicles/available:
    get:
      tags: [Vehicles]
      summary: Get available vehicles
      description: Get all vehicles with AVAILABLE status
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Available vehicles
  /vehicles/{id}/status:
    patch:
      tags: [Vehicles]
      summary: Update vehicle status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/VehicleStatus'
      responses:
        '200':
          description: Status updated

  # ==================== SCHEDULES ====================
...
security:
  - BearerAuth: []